Refactor bindings storage into its own container
