+++ Assets/Script/Input/Bindings/BindingsContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf507de97657c1b4bba8f32848eb4ef2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
